requirements.txt
brian2>=2.5.1
brian2hears>=0.9.2
git+https://github.com/Zakhar-the-Robot/brain_pycore.git
pillow